requirements.txt
pycli-tools>=2.0.2
bottle==0.12.11
Pillow==4.0.0
https://github.com/smarnach/pyexiftool/archive/v0.2.0.tar.gz#egg=pyexiftool-0.1.0
